Move strategies to test subfolder
