+++ tests/strategy/strats/default_strategy.py
