+++ tests/strategy/strats/failing_strategy.py
