+++ tests/strategy/strats/legacy_strategy.py
