+++ tests/strategy/test_interface.py
@@ -11,7 +11,7 @@ from freqtrade.data.converter import parse_ticker_dataframe
 from freqtrade.data.history import load_tickerdata_file
 from freqtrade.persistence import Trade
 from freqtrade.resolvers import StrategyResolver
-from freqtrade.strategy.default_strategy import DefaultStrategy
+from .strats.default_strategy import DefaultStrategy
 from tests.conftest import get_patched_exchange, log_has
 
 # Avoid to reinit the same object again and again
