+++ tests/strategy/test_strategy.py
@@ -31,21 +31,21 @@ def test_search_strategy():
 
 
 def test_search_all_strategies_no_failed():
-    directory = Path(__file__).parent
+    directory = Path(__file__).parent / "strats"
     strategies = StrategyResolver.search_all_objects(directory, enum_failed=False)
     assert isinstance(strategies, list)
-    assert len(strategies) == 3
+    assert len(strategies) == 2
     assert isinstance(strategies[0], dict)
 
 
 def test_search_all_strategies_with_failed():
-    directory = Path(__file__).parent
+    directory = Path(__file__).parent / "strats"
     strategies = StrategyResolver.search_all_objects(directory, enum_failed=True)
     assert isinstance(strategies, list)
-    assert len(strategies) == 4
+    assert len(strategies) == 3
     # with enum_failed=True search_all_objects() shall find 3 good strategies
     # and 1 which fails to load
-    assert len([x for x in strategies if x['class'] is not None]) == 3
+    assert len([x for x in strategies if x['class'] is not None]) == 2
     assert len([x for x in strategies if x['class'] is None]) == 1
 
 
@@ -326,7 +326,7 @@ def test_strategy_override_use_sell_profit_only(caplog, default_conf):
 
 @pytest.mark.filterwarnings("ignore:deprecated")
 def test_deprecate_populate_indicators(result, default_conf):
-    default_location = path.join(path.dirname(path.realpath(__file__)))
+    default_location = Path(__file__).parent / "strats"
     default_conf.update({'strategy': 'TestStrategyLegacy',
                          'strategy_path': default_location})
     strategy = StrategyResolver.load_strategy(default_conf)
@@ -360,7 +360,7 @@ def test_deprecate_populate_indicators(result, default_conf):
 
 @pytest.mark.filterwarnings("ignore:deprecated")
 def test_call_deprecated_function(result, monkeypatch, default_conf):
-    default_location = path.join(path.dirname(path.realpath(__file__)))
+    default_location = Path(__file__).parent / "strats"
     default_conf.update({'strategy': 'TestStrategyLegacy',
                          'strategy_path': default_location})
     strategy = StrategyResolver.load_strategy(default_conf)
